# Improve performance of content size calculation in List