### Description:
Makes calculating the content size in List O(n) in the number of custom-height items rather than the number of total items. Unsurprisingly, a huge performance improvement for long lists with only a few custom item heights (as is a common use case).

Fixes #3427 (for list)

### Checklist:

- [ ] Tests included. <- already covered by existing tests
- [x] Lint and formatter run with no errors.
- [x] Tests all pass.

Benchmark (before):

`goos: darwin
goarch: arm64
pkg: fyne.io/fyne/v2/widget`

`BenchmarkContentMinSize-8   	     198	   6114649 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	fyne.io/fyne/v2/widget	2.062s`

Benchmark (after):

`goos: darwin
goarch: arm64
pkg: fyne.io/fyne/v2/widget`

`BenchmarkContentMinSize-8   	19891040	        51.04 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	fyne.io/fyne/v2/widget	1.565s`